refactor(examples): extract derivatives module
